# [SUP-926] Performance optimization of scanning endpoints