--- a/core/pom.xml
+++ b/core/pom.xml
@@ -40,7 +40,7 @@
     <dependency>
       <groupId>org.mockito</groupId>
       <artifactId>mockito-core</artifactId>
-      <version>3.11.2</version>
+      <version>5.4.0</version>
       <scope>test</scope>
     </dependency>
   </dependencies>