--- a/snyk-sdk/src/main/java/io/snyk/sdk/api/v1/SnykClient.java
+++ b/snyk-sdk/src/main/java/io/snyk/sdk/api/v1/SnykClient.java
@@ -77,6 +77,7 @@ public class SnykClient {
       ))
       .withQueryParam("org", organisation)
       .withQueryParam("repository", repository)
+      .withQueryParam("topLevelOnly", "true")
       .build();
     HttpResponse<String> response = httpClient.send(request, HttpResponse.BodyHandlers.ofString());
     return SnykResult.createResult(response, TestResult.class);
@@ -90,6 +91,7 @@ public class SnykClient {
         URLEncoder.encode(version, UTF_8)
       ))
       .withQueryParam("org", organisation)
+      .withQueryParam("topLevelOnly", "true")
       .build();
     HttpResponse<String> response = httpClient.send(request, HttpResponse.BodyHandlers.ofString());
     return SnykResult.createResult(response, TestResult.class);
@@ -103,6 +105,7 @@ public class SnykClient {
         URLEncoder.encode(version, UTF_8)
       ))
       .withQueryParam("org", organisation)
+      .withQueryParam("topLevelOnly", "true")
       .build();
     HttpResponse<String> response = httpClient.send(request, HttpResponse.BodyHandlers.ofString());
     return SnykResult.createResult(response, TestResult.class);