Currently the `/test` endpoint of the Snyk V1 API builds full dependency graphs for each package it receives. For this plugin, this means that when you do a `pip install mlflow`, you'll wait for the entire transitive dependency line of `mlflow` to be finished before `pip` continues.

The next thing `pip` does is to download the first transitive package of `mlflow`, which gets the same treatment inside Snyk, and thus it continues, wasting large amounts of time.

We recently introduced a fix to Snyk that loosens this logic of the `/test` endpoint and allows for testing of single packages while we wait for a the longer term solution to be completed by Packages and Dependencies and the Package KB project.

### Testing

There already are pretty thorough integration tests validating the `/test` endpoint results, and I don't see unit tests trying to mock that behavior, which is probably fine. Hence no new tests have been added.

### Benchmark:

Doing `pip3 install mlflow`.

Before:
```
12.24s user 2.75s system 3% cpu 7:12.33 total
```

After:
```
12.14s user 2.60s system 34% cpu 42.320 total
```

Almost a 10x performance increase! 🏃🏻 💨